chore: move jetstream components to atoms
