+++ resources/js/Components/CategoryModal.vue
@@ -48,7 +48,7 @@
 </template>
 
 <script>
-    import Modal from '@/Jetstream/Modal.vue'
+    import Modal from '@/Components/atoms/Modal.vue'
     import { useForm } from "@inertiajs/inertia-vue3"
     import { AtField, AtInput, AtButton } from "atmosphere-ui"
     import { reactive, toRefs } from '@vue/reactivity'
