+++ resources/js/Components/ImportResourceModal.vue
@@ -42,7 +42,7 @@
 import { ref } from "vue";
 import { AtButton } from "atmosphere-ui";
 
-import Modal from "@/Jetstream/Modal.vue";
+import Modal from "@/Components/atoms/Modal.vue";
 import ImportHolder from "@/Components/organisms/ImportHolder.vue";
 import TabSelector from "./TabSelector.vue";
 import LogerTabButton from "./atoms/LogerTabButton.vue";
