+++ resources/js/Components/TransactionModal.vue
@@ -172,7 +172,7 @@
 <script setup>
     import { format } from 'date-fns'
     import { reactive, toRefs, watch, computed, inject, ref } from 'vue'
-    import Modal from '@/Jetstream/Modal.vue'
+    import Modal from '@/Components/atoms/Modal.vue'
     import { useForm } from "@inertiajs/inertia-vue3"
     import { AtField, AtButton, AtFieldCheck, AtInput } from "atmosphere-ui"
     import { NSelect, NDatePicker } from "naive-ui";
