+++ resources/js/Components/atoms/ApplicationMark.vue
