+++ resources/js/Components/atoms/AuthenticationCard.vue
