+++ resources/js/Components/atoms/AuthenticationCardLogo.vue
