+++ resources/js/Components/atoms/Banner.vue
