+++ resources/js/Components/atoms/SectionCardTitle.vue
