+++ resources/js/Components/atoms/SectionTitle.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script setup>
-import { computed } from "@vue/runtime-core";
+import { computed } from "vue";
 
 const props = defineProps({
     type: {
