+++ resources/js/Components/templates/AppLayout.vue
@@ -221,10 +221,10 @@
     import { useI18n } from 'vue-i18n'
     import { Link, usePage } from '@inertiajs/inertia-vue3'
 
-    import JetBanner from '@/Jetstream/Banner.vue'
-    import JetDropdown from '@/Jetstream/Dropdown.vue'
-    import JetDropdownLink from '@/Jetstream/DropdownLink.vue'
-    import JetResponsiveNavLink from '@/Jetstream/ResponsiveNavLink.vue'
+    import JetBanner from '@/Components/atoms/Banner.vue'
+    import JetDropdown from '@/Components/atoms/Dropdown.vue'
+    import JetDropdownLink from '@/Components/atoms/DropdownLink.vue'
+    import JetResponsiveNavLink from '@/Components/atoms/ResponsiveNavLink.vue'
     import PrivacyToggle from '@/Components/molecules/PrivacyToggle.vue'
     import LogerTabButton from '@/Components/atoms/LogerTabButton.vue'
     import AppShell from './AppShell.vue'
