+++ resources/js/Pages/Profile/LogoutOtherBrowserSessionsForm.vue
@@ -82,13 +82,13 @@
 </template>
 
 <script>
-    import JetActionMessage from '@/Jetstream/ActionMessage.vue'
-    import JetActionSection from '@/Jetstream/ActionSection.vue'
-    import JetButton from '@/Jetstream/Button.vue'
-    import JetDialogModal from '@/Jetstream/DialogModal.vue'
-    import JetInput from '@/Jetstream/Input.vue'
-    import JetInputError from '@/Jetstream/InputError.vue'
-    import JetSecondaryButton from '@/Jetstream/SecondaryButton.vue'
+    import JetActionMessage from '@/Components/atoms/ActionMessage.vue'
+    import JetActionSection from '@/Components/atoms/ActionSection.vue'
+    import JetButton from '@/Components/atoms/Button.vue'
+    import JetDialogModal from '@/Components/atoms/DialogModal.vue'
+    import JetInput from '@/Components/atoms/Input.vue'
+    import JetInputError from '@/Components/atoms/InputError.vue'
+    import JetSecondaryButton from '@/Components/atoms/SecondaryButton.vue'
 
     export default {
         props: ['sessions'],
