+++ resources/js/Pages/Profile/UpdateProfileInformationForm.vue
@@ -70,13 +70,13 @@
 </template>
 
 <script>
-    import JetButton from '@/Jetstream/Button.vue'
-    import JetFormSection from '@/Jetstream/FormSection.vue'
-    import JetInput from '@/Jetstream/Input.vue'
-    import JetInputError from '@/Jetstream/InputError.vue'
-    import JetLabel from '@/Jetstream/Label.vue'
-    import JetActionMessage from '@/Jetstream/ActionMessage.vue'
-    import JetSecondaryButton from '@/Jetstream/SecondaryButton.vue'
+    import JetButton from '@/Components/atoms/Button.vue'
+    import JetFormSection from '@/Components/atoms/FormSection.vue'
+    import JetInput from '@/Components/atoms/Input.vue'
+    import JetInputError from '@/Components/atoms/InputError.vue'
+    import JetLabel from '@/Components/atoms/Label.vue'
+    import JetActionMessage from '@/Components/atoms/ActionMessage.vue'
+    import JetSecondaryButton from '@/Components/atoms/SecondaryButton.vue'
     import { AtField } from "atmosphere-ui"
     import LogerInput from '@/Components/atoms/LogerInput.vue'
