+++ resources/js/Pages/Teams/CreateTeamForm.vue
@@ -38,11 +38,11 @@
 </template>
 
 <script setup>
-    import JetButton from '@/Jetstream/Button.vue'
-    import JetFormSection from '@/Jetstream/FormSection.vue'
-    import JetInput from '@/Jetstream/Input.vue'
-    import JetInputError from '@/Jetstream/InputError.vue'
-    import JetLabel from '@/Jetstream/Label.vue'
+    import JetButton from '@/Components/atoms/Button.vue'
+    import JetFormSection from '@/Components/atoms/FormSection.vue'
+    import JetInput from '@/Components/atoms/Input.vue'
+    import JetInputError from '@/Components/atoms/InputError.vue'
+    import JetLabel from '@/Components/atoms/Label.vue'
     import { Inertia } from '@inertiajs/inertia'
     import TeamForm from '../Onboarding/TeamForm.vue'
